backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib==1.7.4
pydantic-settings==2.10.1
langgraph==0.6.6
requests==2.32.5
pypdf==5.2.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
google-auth==2.40.3
